extract dataclasses for transport2030
