+++ generatorcore/transport2030/t30.py
@@ -1,0 +1,79 @@
+from dataclasses import dataclass, asdict
+
+from .investmentaction import InvestmentAction
+from .air import Air
+from .road import (
+    Road,
+    RoadBus,
+    RoadCar,
+    RoadGoods,
+    RoadGoodsLightDuty,
+    RoadGoodsMediumAndHeavyDuty,
+    RoadPeople,
+    RoadSum,
+    RoadInvestmentAction,
+)
+from .rail import Rail, RailGoods, RailPeople, RailPeopleMetroActionInfra, RailPeopleSum
+from .ship import Ship, ShipDomestic, ShipInternational, ShipDomesticActionInfra
+from .other import Other, OtherCycle, OtherFoot
+from .transport import Transport
+from .dataclasses import EnergySum, GPlanning, G, T
+
+
+@dataclass
+class T30:
+    air_inter: Transport
+    air_dmstc: Transport
+    road_ppl: RoadPeople
+    road_car: RoadCar
+    road_car_it_ot: Road
+    road_car_ab: Road
+    road_bus: RoadBus
+    road_gds: RoadGoods
+    road_gds_ldt: RoadGoodsLightDuty
+    road_gds_ldt_it_ot: Road
+    road_gds_ldt_ab: Road
+    road_gds_mhd: RoadGoodsMediumAndHeavyDuty
+    road_gds_mhd_it_ot: Road
+    road_gds_mhd_ab: Road
+    road_gds_mhd_action_wire: InvestmentAction
+    rail_ppl: RailPeopleSum
+    rail_ppl_distance: RailPeople
+    rail_ppl_metro: RailPeople
+    rail_gds: RailGoods
+    ship: Ship
+    ship_dmstc: ShipDomestic
+    ship_dmstc_action_infra: ShipDomesticActionInfra
+    ship_inter: ShipInternational
+    other_foot: OtherFoot
+    other_foot_action_infra: InvestmentAction
+    other_cycl: OtherCycle
+    other_cycl_action_infra: InvestmentAction
+    g_planning: GPlanning
+    s: EnergySum
+    s_diesel: EnergySum
+    s_emethan: EnergySum
+    s_jetfuel: EnergySum
+    s_petrol: EnergySum
+    s_fueloil: EnergySum
+    s_lpg: EnergySum
+    s_gas: EnergySum
+    s_biogas: EnergySum
+    s_bioethanol: EnergySum
+    s_biodiesel: EnergySum
+    s_elec: EnergySum
+    s_hydrogen: EnergySum
+    g: G
+    t: T
+    air: Air
+    road: RoadSum
+    rail: Rail
+    other: Other
+    rail_action_invest_infra: InvestmentAction
+    rail_action_invest_station: InvestmentAction
+    rail_ppl_metro_action_infra: RailPeopleMetroActionInfra
+    road_action_charger: RoadInvestmentAction
+    road_bus_action_infra: InvestmentAction
+
+    def dict(self):
+        return asdict(self)
